+++ TabletopTweaks/NewContent/MythicFeats/MythicManyshot.cs
@@ -41,6 +41,7 @@ namespace TabletopTweaks.NewContent.MythicFeats {
         [HarmonyPatch(typeof(RuleAttackWithWeapon), nameof(RuleAttackWithWeapon.LaunchProjectiles), new Type[] { typeof(BlueprintProjectileReference[]) })]
         static class ItemEntityWeapon_HoldInTwoHands_Patch {
             static bool Prefix(RuleAttackWithWeapon __instance, BlueprintProjectileReference[] projectiles) {
+                if (ModSettings.AddedContent.MythicFeats.IsDisabled("MythicManyshot")) { return true; }
                 foreach (BlueprintProjectileReference blueprintProjectileReference in projectiles) {
                     if (blueprintProjectileReference.Get() != null) {
                         __instance.LaunchProjectile(blueprintProjectileReference.Get(), true);